Update ManagementService.py
as I wrote yesterday: chaining the single services works, but using the (pedestrian) Management service directly does not. Here the error that gets thrown:

  File "/home/ssteinig/wypmodels/Network-PedestrianOnly/bottle.py", line 1575, in wrapper
    rv = callback(*a, **ka)
  File "ManagementService.py", line 140, in service
    return management(start_point, walking_time_period, walking_speed, distance_decay_function)
  File "ManagementService.py", line 128, in management
    result = '{"walkshed": %s, "poi": %s}' % (aggregation_data)
TypeError: not enough arguments for format string

which seems to make sense, because in the Bike Management service it looks like this:
result = '{"walkshed": %s, "poi": %s}' % (aggregation_data, poi_data)

So when I added point_data it runs.